requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
jinja2
aiofiles
